+++ CHANGELOG.md
@@ -1,17 +1,33 @@
-## 0.0.1 (2022-04-26)
+## 0.0.1 (2022-06-24)
 
 
 ### ✨ Features | 新功能
 
+* 输出产物配置化 ([8ce7d10](https://github.com/alqmc/build/commit/8ce7d10))
+* build-vue ([99a15af](https://github.com/alqmc/build/commit/99a15af))
+* built-ts build-utils ([8f71a0d](https://github.com/alqmc/build/commit/8f71a0d))
+* built-ts build-utils ([319e4a5](https://github.com/alqmc/build/commit/319e4a5))
 * dependencies ([e6ec46c](https://github.com/alqmc/build/commit/e6ec46c))
+* detail ([6d89c3a](https://github.com/alqmc/build/commit/6d89c3a))
+* detail ([12d4586](https://github.com/alqmc/build/commit/12d4586))
 * init ([6265ad0](https://github.com/alqmc/build/commit/6265ad0))
+* vue ([ee663d1](https://github.com/alqmc/build/commit/ee663d1))
+* vue lib ([f8a217f](https://github.com/alqmc/build/commit/f8a217f))
+* vue lib ([21575af](https://github.com/alqmc/build/commit/21575af))
 
 
 ### 🐛 Bug Fixes | Bug 修复
 
 * deatil ([91e5098](https://github.com/alqmc/build/commit/91e5098))
+* detail ([b3f51e1](https://github.com/alqmc/build/commit/b3f51e1))
 * eslint ([8365472](https://github.com/alqmc/build/commit/8365472))
 * log ([88cf3a6](https://github.com/alqmc/build/commit/88cf3a6))
+* merge rollup outoptions ([f45a4f4](https://github.com/alqmc/build/commit/f45a4f4))
+
+
+### 📝 Documentation | 文档
+
+* log ([cb1a3f8](https://github.com/alqmc/build/commit/cb1a3f8))